Rename project TagsCloud.Visualization to TagsCloud
