+++ cs/TagsCloud.Core/Properties/FriendlyAssemblies.cs
@@ -1,4 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("TagsCloud.Tests")]
-[assembly: InternalsVisibleTo("TagsCloud.Visualization")]
+[assembly: InternalsVisibleTo("TagsCloud")]
